Extract getSlice and getUseHook helpers
